将hint及format方法抽出，摆脱对config模块的依赖
