+++ lib/config.js
@@ -8,6 +8,7 @@ var path = require('path');
 var stripJsonComments = require('strip-json-comments');
 
 var util = require('./util');
+var fsUtil = require('./fs-util');
 
 /**
  * Name of the config file.
@@ -59,7 +60,7 @@ var getConfigFileUnderOrUp = function (dirPath) {
  * @return {?string} path of the config file, null if not exists
  */
 var getConfigFileUnderHomePath = function () {
-    var homePath = util.getHomePath();
+    var homePath = fsUtil.getHomePath();
     return homePath ? getConfigFileUnder(homePath) : null;
 };
 
@@ -85,7 +86,7 @@ var getConfigFile = util.cachable(function (dirPath) {
     }
 
     // use default config file
-    return path.join(util.app.root, 'lib/default/htmlcsrc');
+    return path.join(fsUtil.app.root, 'lib/default/htmlcsrc');
 });
 
 /**
