+++ lib/util.js
@@ -3,19 +3,7 @@
  * @author nighca<nighca@live.cn>
  */
 
-var fs = require('fs');
-var path = require('path');
 var ElementType = require('domelementtype');
-
-/**
- * Application (htmlcs) info.
- *
- * @type {Object}
- * @property {string} root - root path of application (htmlcs) code
- */
-var app = {
-    root: path.resolve(__dirname, '../')
-};
 
 /**
  * Copy properties from src to target.
@@ -69,29 +57,6 @@ var cachable = function (getter) {
     };
 
     return extend(get, {clear: clear});
-};
-
-/**
- * Get path of home(~).
- *
- * @return {string} path of home
- */
-var getHomePath = function () {
-    var homePath = '';
-    var environment = process.env;
-    var paths = [
-        environment.USERPROFILE,
-        environment.HOME,
-        environment.HOMEPATH,
-        environment.HOMEDRIVE + environment.HOMEPATH
-    ];
-
-    while (paths.length) {
-        homePath = paths.shift();
-        if (fs.existsSync(homePath)) {
-            return homePath;
-        }
-    }
 };
 
 /**
@@ -177,13 +142,11 @@ var extractCommentInfo = function (comment) {
 };
 
 module.exports = {
-    app: app,
     nodeType: ElementType,
     isElement: ElementType.isTag,
     extend: extend,
     extendAttribute: extendAttribute,
     cachable: cachable,
-    getHomePath: getHomePath,
     getPosition: getPosition,
     extractCommentInfo: extractCommentInfo
 };
