+++ test/lib/config.spec.js
@@ -6,7 +6,7 @@
 var sh = require('shelljs');
 var fs = require('fs');
 var path = require('path');
-var util = require('../../lib/util');
+var fsUtil = require('../../lib/fs-util');
 var config = require('../../lib/config');
 
 describe('load', function () {
@@ -60,7 +60,7 @@ describe('load', function () {
 
     describe('config file in home path', function () {
         it('should return right config', function () {
-            createConfigFile(util.getHomePath());
+            createConfigFile(fsUtil.getHomePath());
 
             expect(config.load(__filename, true).test).toBe(true);
