+++ test/lib/util.spec.js
@@ -3,23 +3,7 @@
  * @author nighca<nighca@live.cn>
  */
 
-var path = require('path');
 var util = require('../../lib/util');
-var packageInfo = require('../../package.json');
-
-describe('app', function () {
-    var app = util.app;
-
-    describe('root', function () {
-        it('should be a string', function () {
-            expect(typeof app.root).toBe('string');
-        });
-
-        it('should be the dir path where package.json locates', function () {
-            expect(require(path.join(app.root, 'package.json'))).toBe(packageInfo);
-        });
-    });
-});
 
 describe('nodeType', function () {
     var nodeType = util.nodeType;
@@ -150,15 +134,6 @@ describe('cachable', function () {
         });
     });
 });
-
-describe('getHomePath', function () {
-    var getHomePath = util.getHomePath;
-
-    it('should return a path', function () {
-        expect(typeof getHomePath()).toBe('string');
-    });
-});
-
 
 describe('getPosition', function () {
     var getPosition = util.getPosition;
